requirements.txt
discord.py
pymongo
dnspython
Pillow
aiohttp
requests
python-dotenv
motor
